+++ sdk-js-web/src/index.ts
@@ -147,4 +147,4 @@ export {
   Segmenter,
 } from "./tasks/segmenter.js";
 
-export const VERSION: string = "0.6.0";
+export const VERSION: string = "0.6.1";